Memory optimization.

+++ GlSplatRendererLib/GlSplatRenderer.h
@@ -6,12 +6,12 @@
 class GlSplatRenderer : public GlSplat::SplatRenderer
 {
 public:
-    GlSplatRenderer(const QVector<Vec3d> &pts, const QVector<Vec3d> &ns,
+    GlSplatRenderer(const QVector<Vec3f> &pts, const QVector<Vec3f> &ns,
                     double radius = 0.01, Vec4d color = Vec4d(1.0, 1.0, 1.0, 1.0));
 
 public:
-    QVector<Vec3d> points;
-    QVector<Vec3d> normals;
+    QVector<Vec3f> points;
+    QVector<Vec3f> normals;
     double mRadius;
     double mColor[4];
